pls work

+++ app/controllers/brands_controller.rb
@@ -82,8 +82,9 @@ class BrandsController < ApplicationController
     
 
     @brand = Brand.new(brand_params.except(:tags))
-    create_or_delete_brands_tags(@brand, params[:brand][:tags])
-    
+    if @auth
+      create_or_delete_brands_tags(@brand, params[:brand][:tags])
+    end
     @brand.brand_text = "#654321;"
     @brand.user = current_user
     delete_styles(@brand, params[:brand][:styles],)
@@ -110,8 +111,9 @@ class BrandsController < ApplicationController
   # PATCH/PUT /brands/1 or /brands/1.json
   def update
     create_or_delete_brands_tags(@brand, params[:brand][:tags])
-    delete_styles(@brand, params[:brand][:styles],)
-
+    if @auth
+      delete_styles(@brand, params[:brand][:styles],)
+    end
 
  
     respond_to do |format|
@@ -162,12 +164,12 @@ class BrandsController < ApplicationController
 
     def create_or_delete_brands_tags(brand, tags)
       brand.brand_tag_assocs.destroy_all
-      if not tags.nil?
-        tags = tags.strip.split(',')
-        tags.each do |tag|
-          brand.tags << Tag.find_or_create_by(name: tag)
-        end
+    
+      tags = tags.strip.split(',')
+      tags.each do |tag|
+        brand.tags << Tag.find_or_create_by(name: tag)
       end
+
     end
